Thumbnails for GIFs/Videos, Alt Text, Giveaway disclaimers
When a GIF or video is posted by the canes, this will now be represented by a static thumbnail of said GIF or video.

Images that have alt text will now have that reflected in the bot.

Some giveaways posted by the canes are invalid on Bluesky. I have tried to handle this by appending a disclaimer to text with particularly indicative keywords.