Video Support!
Big update. Videos can now be posted! The code detects when a url is a video. If the video is too long for Bluesky or the amount of daily videos has been exceeded, the bot will post a thumbnail instead. Videos will scale based on their aspect ratio. Some possible kinks to iron out, I was having an issue with getting rate limited that I believe I have solved but will still be monitoring. Would ideally like to move to a single agent for posting, but this should work in the meantime.