Bug fixes
A very nice person told me that my app password was leaked in my codebase. This has been remedied, and the app password that was present is no longer valid to access the account. They also told me that my cron needed to be in UTC, so hopefully this new cron does the trick.